Create functions for playing and displaying game
I've created the playRound function that will play one round of war and
use recursion when a war happens. Also created a function to update the
card totals for the player and computer. Game class has been updated to
use a singleton pattern to track the decks and give access for functions