Refactor Game class to be more object oriented
I've changed the Game class to not have its variables be public and
added getters and setters. I also moved the playRound function into the
Game class because I think it makes more sense to be there since the
game consists of multiple rounds. I fixed a bug around cards drawn for
a war being lost when multiple wars happen in a row.